+++ clients/unshaded/src/main/java/tachyon/client/next/file/TachyonFile.java
@@ -19,13 +19,13 @@ package tachyon.client.next.file;
  * A file handler for a file in Tachyon. It is just a wrapper around the file ID for now.
  */
 public class TachyonFile {
-  private final int mFileId;
+  private final long mFileId;
 
-  public TachyonFile(int fileId) {
+  public TachyonFile(long fileId) {
     mFileId = fileId;
   }
 
-  public int getFileId() {
+  public long getFileId() {
     return mFileId;
   }
 }
